+++ src/sound/YMF278.cc
@@ -25,7 +25,7 @@ static const unsigned EG_TIMER_OVERFLOW = 1 << EG_SH;
 // envelope output entries
 static const int ENV_BITS      = 10;
 static const int ENV_LEN       = 1 << ENV_BITS;
-static const double ENV_STEP   = 128.0 / ENV_LEN;
+static const float ENV_STEP    = 128.0f / ENV_LEN;
 static const int MAX_ATT_INDEX = (1 << (ENV_BITS - 1)) - 1; // 511
 static const int MIN_ATT_INDEX = 0;
 
@@ -54,7 +54,7 @@ static const int mix_level[8] = {
 
 // decay level table (3dB per step)
 // 0 - 15: 0, 3, 6, 9,12,15,18,21,24,27,30,33,36,39,42,93 (dB)
-#define SC(db) unsigned(db * (2.0 / ENV_STEP))
+#define SC(db) unsigned(db * (2.0f / ENV_STEP))
 static const unsigned dl_tab[16] = {
  SC( 0), SC( 1), SC( 2), SC(3 ), SC(4 ), SC(5 ), SC(6 ), SC( 7),
  SC( 8), SC( 9), SC(10), SC(11), SC(12), SC(13), SC(14), SC(31)
@@ -134,24 +134,24 @@ static const byte eg_rate_shift[64] = {
 // TODO check if frequency matches real chip
 #define O(a) int((EG_TIMER_OVERFLOW / a) / 6)
 static const int lfo_period[8] = {
-	O(0.168), O(2.019), O(3.196), O(4.206),
-	O(5.215), O(5.888), O(6.224), O(7.066)
+	O(0.168f), O(2.019f), O(3.196f), O(4.206f),
+	O(5.215f), O(5.888f), O(6.224f), O(7.066f)
 };
 #undef O
 
 
 #define O(a) int(a * 65536)
 static const int vib_depth[8] = {
-	O(0),	   O(3.378),  O(5.065),  O(6.750),
-	O(10.114), O(20.170), O(40.106), O(79.307)
+	O( 0.0f  ), O( 3.378f), O( 5.065f), O( 6.750f),
+	O(10.114f), O(20.170f), O(40.106f), O(79.307f)
 };
 #undef O
 
 
-#define SC(db) int(db * (2.0 / ENV_STEP))
+#define SC(db) int(db * (2.0f / ENV_STEP))
 static const int am_depth[8] = {
-	SC(0),	   SC(1.781), SC(2.906), SC(3.656),
-	SC(4.406), SC(5.906), SC(7.406), SC(11.91)
+	SC(0.0f  ), SC(1.781f), SC(2.906f), SC( 3.656f),
+	SC(4.406f), SC(5.906f), SC(7.406f), SC(11.91f )
 };
 #undef SC
